packages/mewkit/src/migrate/modules/codex/root/.agents/skills/multimodal/scripts/requirements.txt
# mk:multimodal dependencies
# Python 3.10+ required

# Google Gemini API (required)
google-genai>=1.0.0

# Image processing (required)
Pillow>=10.0.0

# Environment variable management (optional — for .env file support)
python-dotenv>=1.0.0
